% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERM_genera.R
\name{ERM_genera}
\alias{ERM_genera}
\title{the function to generate the Environment Relationship Matrix}
\usage{
ERM_genera(text)
}
\arguments{
\item{text}{the fam file to input, can get from the fam_genera()}
}
\value{
matrix_family the ERM of family, for the file with ".grm.gz"

matrix_family_id the file with ".grm.id

matrix_couple the ERM of couple, for the file with ".grm.gz"

matrix_couple_id the file with ".grm.id

matrix_fullsilb the ERM of fullsilbing, for the file with ".grm.gz"

matrix_fullsilb the file with ".grm.id
}
\description{
generate the environment relationship matrix from three aspects: family, couple, fullsilbings.
}
\examples{
\dontrun{
text=fam_genera(100)
l=ERM_genera(text)
}
}
\references{
Yang J, Lee SH, Goddard ME, Visscher PM (2011) GCTA: a tool for genome-wide complex trait analy- sis. Am J Hum Genet 88: 76–82. doi: 10.1016/j.ajhg.2010.11.011 PMID: 21167468

Xia, C. et al. Pedigree- and SNP-associated genetics and recent environment are the major contributors to anthropometric and cardiometabolic trait variation. PLoS. Genet. 12, e1005804 (2016).
}
