% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam_genera.R
\name{fam_genera}
\alias{fam_genera}
\title{the function generate fam file for families}
\usage{
fam_genera(n)
}
\arguments{
\item{n}{the number of families, this is an integer.}
}
\value{
the fam file, it is a data.frame output.
}
\description{
the function to generate the fam file of families containing two parents and two children.
}
\details{
This function is to generate the .fam file for PLINK use only after adding the information of genotypes and phenotypes. And here it sets the structures of the families already. For each family, there are two parents, and two siblings.
}
\examples{
\dontrun{
fam_genera(100)
}
}
