% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam_relation.R
\name{fam_relation}
\alias{fam_relation}
\title{the function generate fam relationship file for families}
\usage{
fam_relation(fam)
}
\arguments{
\item{fam}{the family file, it is a data.frame here.}
}
\value{
the fam_relationship, this is a data.frame output.
}
\description{
the function to generate the fam relationship file of families containing two parents and two children.
}
\details{
Here, the fam relationship is created by the fam file from fam_genera(). And the relationship would be "couple","parentOffspring" or "fullSib".
}
\examples{
\dontrun{
fam <- fam_genera(100)
fam_relation(fam)
}
}
