% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_genera.R
\name{geno_genera}
\alias{geno_genera}
\title{the function generate genotype file}
\usage{
geno_genera(text, m)
}
\arguments{
\item{text}{the fam file, it is a data.frame here.}

\item{m}{the number of SNPs, it is an integer.}
}
\value{
the genotype file, the output is a data.frame.
}
\description{
the function to generate the genotype file of families containing two parents and two children.
}
\examples{
\dontrun{
text <- fam_genera(50)
m <- 100
gene_match(text,m)
}
}
