% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{meanC}
\alias{meanC}
\title{A mean function using Rcpp}
\usage{
meanC(x)
}
\arguments{
\item{x}{the vector to get the mean}
}
\value{
the mean of x
}
\description{
A mean function using Rcpp
}
\examples{
\dontrun{
meanC(c(1,2,3))
}
}
