% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_MetropolisR.R
\name{rw.Metropolis}
\alias{rw.Metropolis}
\title{Random walk Metropolis sampler for generating the standard Laplace distribution}
\usage{
rw.Metropolis(sigma, x0, N)
}
\arguments{
\item{sigma}{the standard deviance of state \code{X_t} of the random walk}

\item{x0}{the initial state of the random walk}

\item{N}{the number of samples}
}
\value{
random samples of size \code{N}
}
\description{
R function to implement a random walk Metropolis sampler for generating the standard Laplace distribution.
}
\examples{
\dontrun{
rw.Metropolis(4,1,100)
}
}
