% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwhishartR.R
\name{rwishart}
\alias{rwishart}
\title{sampler for Wishart distribution}
\usage{
rwishart(n, sigma)
}
\arguments{
\item{n}{the scale parameter of Wishart distribution}

\item{sigma}{the mean matrix of Wishart distribution}
}
\value{
a random sample from Wishart distribution
}
\description{
R function to implement a sampler for Wishart distribution.
}
\examples{
\dontrun{
sigma=matrix(c(5,1,1,3),2,2)
rwishart(5,sigma)
}
}
