\name{SC19036-package}
\alias{SC19036-package}
\docType{package}
\title{
  The R Package for StatComp
}
\description{
  The R package is for statistics computing curriculum containing all the homework answers from the course by Hu Die.
}
\details{
  The main part of the package is the homework answers for each course by Hu Die. And the functions in the homework would be listed with the details of each function.
}
\author{
Hu Die

Maintainer: Hu Die <janehoe@mail.ustc.edu.cn>
}